Make prerender async.
